Add files via upload
A beginner project stimulated on Wokwi. It uses an LDR sensor to automatically dim the LED that acts as the headlight. The potentiometer is used as a knob for the left/right indicator. The DHT22 sensor is used to measure the temperature and humidity outside. Finally, all of the information is projected on the LCD screen.

The next improvement would be to add constraints to control the auto dimming according to fog levels and to add radar to initiate partial dimming of the headlights when vehicles travel in opposite directions.